+++ TerrainSDK/vtdata/ElevationGrid.h
@@ -85,7 +85,7 @@ public:
 	// Set/Get height values
 	void  SetFValue(int i, int j, float value);
 	void  SetValue(int i, int j, short value);
-	int   GetValue(int i, int j) const;	// returns height value as a integer
+	short GetValue(int i, int j) const;	// returns height value as a integer
 	float GetFValue(int i, int j) const;	// returns height value as a float
 	float GetFValueSafe(int i, int j) const;
